Combine scripts.
